Separate Windows routing implementation from Unix
To better maintain both codebases, separate out Windows' routing manager from
the unix implementations to ensure correctness.
