+++ talpid-core/src/routing/windows.rs
@@ -1,6 +1,5 @@
 use super::NetNode;
 use crate::winnet;
-use futures::{sync::oneshot, Async, Future};
 use ipnetwork::IpNetwork;
 use std::collections::HashMap;
 
@@ -14,16 +13,16 @@ pub enum Error {
 
 pub type Result<T> = std::result::Result<T, Error>;
 
-pub struct RouteManagerImpl {
-    shutdown_rx: oneshot::Receiver<oneshot::Sender<()>>,
-    is_manager_shut_down: bool,
+/// Manages routes by calling into WinNet
+pub struct RouteManager {
+    callback_handles: Vec<winnet::WinNetCallbackHandle>,
+    is_stopped: bool,
 }
 
-impl RouteManagerImpl {
-    pub fn new(
-        required_routes: HashMap<IpNetwork, NetNode>,
-        shutdown_rx: oneshot::Receiver<oneshot::Sender<()>>,
-    ) -> Result<Self> {
+impl RouteManager {
+    /// Creates a new route manager that will apply the provided routes and ensure they exist until
+    /// it's stopped.
+    pub fn new(required_routes: HashMap<IpNetwork, NetNode>) -> Result<Self> {
         let routes: Vec<_> = required_routes
             .iter()
             .map(|(destination, node)| {
@@ -41,41 +40,47 @@ impl RouteManagerImpl {
             return Err(Error::FailedToStartManager);
         }
 
-
         Ok(Self {
-            shutdown_rx,
-            is_manager_shut_down: false,
+            callback_handles: vec![],
+            is_stopped: false,
         })
     }
 
-    fn shutdown(&mut self) {
-        if !self.is_manager_shut_down {
-            winnet::deactivate_routing_manager();
-            self.is_manager_shut_down = true;
+    /// Sets a callback that is called whenever the default route changes.
+    #[cfg(target_os = "windows")]
+    pub fn add_default_route_callback<T: 'static>(
+        &mut self,
+        callback: Option<winnet::DefaultRouteChangedCallback>,
+        context: T,
+    ) {
+        if self.is_stopped {
+            return;
         }
-    }
-}
 
-impl Drop for RouteManagerImpl {
-    fn drop(&mut self) {
-        self.shutdown();
-    }
-}
-
-impl Future for RouteManagerImpl {
-    type Item = ();
-    type Error = Error;
-    fn poll(&mut self) -> Result<Async<()>> {
-        match self.shutdown_rx.poll() {
-            Ok(Async::Ready(result_tx)) => {
-                self.shutdown();
-                if let Err(_e) = result_tx.send(()) {
-                    log::error!("Receiver already down");
-                }
-                Ok(Async::Ready(()))
+        match winnet::add_default_route_change_callback(callback, context) {
+            Err(_e) => {
+                // not sure if this should panic
+                log::error!("Failed to add callback!");
             }
-            Ok(Async::NotReady) => Ok(Async::NotReady),
-            Err(_) => Ok(Async::Ready(())),
+            Ok(handle) => {
+                self.callback_handles.push(handle);
+            }
         }
+    }
+
+    /// Stops the routing manager and invalidates the route manager - no new default route callbacks
+    /// can be added
+    pub fn stop(&mut self) {
+        if !self.is_stopped {
+            self.callback_handles.clear();
+            winnet::deactivate_routing_manager();
+            self.is_stopped = true;
+        }
+    }
+}
+
+impl Drop for RouteManager {
+    fn drop(&mut self) {
+        self.stop();
     }
 }
